Do not run collectstatic or compress in Debian
The Debian packages are handling compress & collectstatic in a
different way as for Ubuntu. This patch avoids running them,
and offload that works to the package. The original collectstatic
indicated that it was a workaround for Debian. See
I813b5f6067bb6ecce279cab7278d9227c4d31d28. Because of this, the
logic around the collectstatic exec has been limited to only
RedHat packaging.

Note that the variable os_package_type can be defined if we're
using Debian packages on top of Ubuntu (this is achieved by
installing the puppet-openstack-debian-fact package on Ubuntu).
This has been discussed in Tokyo with Emilien.

Also, this package fixes the configuration file path for Apache.

Change-Id: I3fbb1b576dd61ddd92921e9f12edfe5df21fddd0
Co-Authored-By: Alex Schultz <aschultz@mirantis.com>